Added random delay between Mojang requests
Mojang seens to give a timeout earlier when there is a one second delay
between requests. Here
https://github.com/UPSman123/MC-name-gen/issues/2#issuecomment-900431753
I suggest that this may be because Mojang recognises the regularity of
our requests.

This random delay is a test to see it this is true. Sadly you still get
a timeout after around 190 requests.